+++ flake.nix
@@ -2,15 +2,33 @@
   inputs = {
     nixpkgs.url = "nixpkgs/nixos-unstable";
 
-    dream2nix.url = "github:nix-community/dream2nix/legacy";
-    dream2nix.inputs.nixpkgs.follows = "nixpkgs";
+    crane.url = "github:ipetkov/crane";
+    crane.inputs.nixpkgs.follows = "nixpkgs";
+
+    flake-utils.url = "github:numtide/flake-utils";
   };
 
-  outputs = { self, dream2nix, ... }:
-    dream2nix.lib.makeFlakeOutputs {
-      systems = ["x86_64-linux"];
-      config.projectRoot = ./.;
-      source = ./.;
-      projects = ./projects.toml;
-    };
+  outputs = { nixpkgs, crane, flake-utils, ... }:
+    flake-utils.lib.eachDefaultSystem (system:
+      let
+        pkgs = nixpkgs.legacyPackages.${system};
+        craneLib = crane.lib.${system};
+      in
+      {
+        packages.default = craneLib.buildPackage {
+          src = craneLib.cleanCargoSource (craneLib.path ./.);
+        };
+
+        devShells.default = craneLib.devShell {
+          # --- Environment variables ----
+          # Needed for rust-analyzer
+          RUST_SRC_PATH = "${pkgs.rust.packages.stable.rustPlatform.rustLibSrc}";
+
+          # --- Extra packages ---
+          packages = [
+            pkgs.bashInteractive
+            pkgs.yt-dlp
+          ];
+        };
+      });
 }